Move internal packages to internal subdir
Update project layout to match other projects that I maintain.
